requirements.txt
kafka-python,
boto3,
fastapi,
uvicorn,
psycopg2 # Install other db drivers if needed